Support multiple directories in g:minisnip_dir
The specific reason I added this is because I wanted to add support for
minisnip in vim-go (it already supports ultisnips and neosnippet).

However, with just a single directory I can't add the vim-go snippet
directory.

This converts g:minisnip_dir to a colon-separated list of paths, like
the PATH environment variable.

There are several ways to do this; I just chose this as the easiest and
most backwards compatible way.