keep the register clean
`keeppatterns` only assures, that the highlighting is not affected.
Therefore it is necessary to delete the last entry in the search history
when performing a search.

Furthermore the value of the `@/` value has to be restored.

Additionally the unnamed register is not corrupted anymore when the
placeholder gets replaced with some value